We're always >=py36 now
