+++ hypothesis-python/tests/cover/test_lookup.py
@@ -22,6 +22,7 @@ import io
 import string
 import sys
 import typing
+from inspect import signature
 from numbers import Real
 
 import pytest
@@ -679,3 +680,53 @@ def test_compat_get_type_hints_aware_of_None_default():
 
     assert typing.get_type_hints(constructor)["a"] == typing.Optional[str]
     assert inspect.signature(constructor).parameters["a"].annotation == str
+
+
+_ValueType = typing.TypeVar("_ValueType")
+
+
+class Wrapper(typing.Generic[_ValueType]):
+    _inner_value: _ValueType
+
+    def __init__(self, inner_value: _ValueType) -> None:
+        self._inner_value = inner_value
+
+
+@given(st.builds(Wrapper))
+def test_issue_2603_regression(built):
+    """It was impossible to build annotated classes with constructors."""
+    assert isinstance(built, Wrapper)
+
+
+class AnnotatedConstructor(typing.Generic[_ValueType]):
+    value: _ValueType  # the same name we have in `__init__`
+
+    def __init__(self, value: int) -> None:
+        """By this example we show, that ``int`` is more important than ``_ValueType``."""
+        assert isinstance(value, int)
+
+
+@given(st.data())
+def test_constructor_is_more_important(data):
+    """Constructor types should take precedence over all other annotations."""
+    data.draw(st.builds(AnnotatedConstructor))
+
+
+def use_signature(self, value: str) -> None:
+    ...
+
+
+class AnnotatedConstructorWithSignature(typing.Generic[_ValueType]):
+    value: _ValueType  # the same name we have in `__init__`
+
+    __signature__ = signature(use_signature)
+
+    def __init__(self, value: int) -> None:
+        """By this example we show, that ``__signature__`` is the most important source."""
+        assert isinstance(value, str)
+
+
+@given(st.data())
+def test_signature_is_the_most_important_source(data):
+    """Signature types should take precedence over all other annotations."""
+    data.draw(st.builds(AnnotatedConstructorWithSignature))
